don't upcast for floats

+++ python/test/unit/language/test_core.py
@@ -2464,9 +2464,9 @@ def test_sum_dtype():
     kernel_default_int[(1, )](out)
     assert out[0] == 32 * 32
 
-    out = torch.empty(1, dtype=torch.float32, device='cuda')
+    out = torch.empty(1, dtype=torch.bfloat16, device='cuda')
     kernel_default_float[(1, )](out)
-    torch.testing.assert_close(out[0], torch.tensor(32 * 32, dtype=torch.float32, device='cuda'))
+    torch.testing.assert_close(out[0], torch.tensor(32 * 32, dtype=torch.bfloat16, device='cuda'))
 
 
 @triton.jit
